Refactor to make RocksDBColumnarKeyValueStorage abstract and extend it to optimistic and pessimistic
